recruitment_app/requirements_all.txt
apturl==0.5.2
asgiref==3.5.0
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cached-property==1.5.2
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.13
distro==1.7.0
distro-info===1.1build1
Django==3.2.12
django-cors-headers==3.13.0
django-filter==22.1
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
exceptiongroup==1.0.1
fasteners==0.14.1
future==0.18.2
gyp==0.1
h11==0.13.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
jwcrypto==1.4.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.0.1
mock==4.0.3
monotonic==1.6
more-itertools==8.10.0
mysqlclient==1.4.6
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
openapi-codec==1.3.2
packaging==21.3
paramiko==2.9.3
pbr==5.8.0
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.0.0
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.2.0
pytest-django==4.5.2
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-xlib==0.29
pytz==2022.1
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
QtPy==2.0.0
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.2
systemd-python==234
texttable==1.6.4
tomli==2.0.1
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.15.0
wadllib==1.3.6
websocket-client==1.2.3
wrapt==1.14.1
wsproto==1.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
